Refactor protocol.rs code generation into its own file
